requirements.txt
Django==3.2
django-active-link==0.1.8
django-allauth==0.47.0
django-cors-headers==3.11.0
django-email-utils==1.0.0
django-filter==21.1
django-model-utils==4.2.0
django-smartfields==1.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-writable-nested==0.6.3